requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-Limiter==3.12
Flask-WTF==1.2.2
WTForms==3.2.1
cryptography==45.0.6
bcrypt==4.3.0
PyJWT==2.10.1
pyotp==2.9.0
qrcode==8.2
python-dotenv==1.1.1
email-validator==2.3.0
itsdangerous==2.2.0
Flask-Mail==0.9.1
